fix: Keep module state when editing students
This addresses a bug pointed out by @blaukc where if we edit a student using the `edit` command and don't intend to change their modules, their module state gets completely wiped.

When we parse the `EditCommand` we create an `EditStudentDescriptor` instance which by default is initialized as a List with 0 entries.

So when executing the command we use the `Optional` `getModules` function to decide if we want to keep the old module state or replace it with new module state that we provided in the command. The bug was that the logic for the for `getModules` always returns a non-null `Optional` value (because it checks if `modules` is null which never happens bcos always initialised with empty list), thereby always taking the empty list of modules from the `EditStudentDescriptor` instance and not the old `Student` instance, thus wiping out the modules state.

This commit initializes `modules` as `null` so that `getModules` works as intended.